Parallel explicit solver

+++ Implicit-Lagrange/Topology_Optimization/Bounded_Strain_Constraint.h
@@ -105,7 +105,7 @@ private:
 
 public:
   bool nodal_density_flag_;
-  size_t last_comm_step, current_step, last_solve_step;
+  int last_comm_step, current_step, last_solve_step;
   std::string my_fea_module = "Elasticity";
 
   BoundedStrainConstraint_TopOpt(FEA_Module *FEM, bool nodal_density_flag, real_t maximum_strain)
