+++ Implicit-Lagrange/Topology_Optimization/Heat_Capacity_Potential_Constraint.h
@@ -110,7 +110,7 @@ private:
 
 public:
   bool nodal_density_flag_;
-  size_t last_comm_step, current_step, last_solve_step;
+  int last_comm_step, current_step, last_solve_step;
   std::string my_fea_module = "Heat_Conduction";
 
   HeatCapacityPotentialConstraint_TopOpt(FEA_Module *FEM, bool nodal_density_flag, real_t constraint_value=0, bool inequality_flag=true)
